Scope detail-view tables to their tab to prevent URL-sort leakage
The TanStack "Column with id 'X' does not exist" warning fires because
each detail view called its lines table's useNonPaginatedMaterialTable
hook in the parent body, so the hook stayed alive on every tab. When
the user (or the DetailTabs onChange) wrote a new sort id to the shared
URL `sort` param, the still-mounted hook handed that id to MRT, whose
columns didn't include it.

Instead of filtering the bad sort id out at the edge (the previous
attempt in 1ce85cd / d4a348b), move each lines table — plus its footer
and line-edit modals — into a DetailsTab component that only mounts
inside the Details TabPanel. Other tabs already follow this pattern
(DocumentsTable, ActivityLogList), so the Details tab was the odd one
out.

Done for InboundShipment and RequestRequisition; other detail views
(supplier-return, response-requisition) likely need the same treatment.

Reverts the validSorting filter added in d4a348b.
